Bye bye gsap :rocket:
