+++ _css/profile-introduction.scss
@@ -15,6 +15,16 @@
     left: 0;
     background-color: #3F51B5;
     opacity: 0;
+
+    &.show {
+      animation: show-rendering-surface 0.5s linear;
+      animation-fill-mode: forwards;
+    }
+
+    @keyframes show-rendering-surface {
+      from {opacity: 0;}
+      to {opacity: 1;}
+    }
   }
 
   .center-content {
