+++ _includes/dependencies-js-home.html
@@ -1,3 +1,3 @@
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/cookieconsent2/3.0.3/cookieconsent.min.js"></script>
-<script type="text/javascript" src="/assets/js/index.home.min.js?rev=4e02995ccb1fee007aef57c09aa6f02e"></script>
+<script type="text/javascript" src="/assets/js/index.home.min.js?rev=9a9d8152939d79b7c979b764ce9c674b"></script>
 {%include service-worker.html %}
