+++ _includes/service-worker-css-home-urls.js
@@ -1,1 +1,1 @@
-"/assets/styles/style.home.css?rev=3b8973e93a74125dc6e26eb376c4b287",
+"/assets/styles/style.home.css?rev=0d10a6f2ee4faf34089fc8f22257e7f3",
