+++ _includes/service-worker-js-home-urls.js
@@ -1,1 +1,1 @@
-"/assets/js/index.home.min.js?rev=765d9647ef7d53499c5079ea5f9ad954",
+"/assets/js/index.home.min.js?rev=4e02995ccb1fee007aef57c09aa6f02e",
