+++ _js/common/animation.js
@@ -1,0 +1,11 @@
+/* @flow */
+import { addCssClass } from '../common/css-class'
+
+const animation = (elementId: string, animation: string): void => {
+  const element = document.getElementById(elementId)
+  if (element) {
+    addCssClass(element, animation)
+  }
+}
+
+export { animation }
