+++ _jsbuild/common/animation.js
@@ -1,0 +1,11 @@
+/*  */
+import { addCssClass } from '../common/css-class'
+
+const animation = (elementId, animation) => {
+  const element = document.getElementById(elementId)
+  if (element) {
+    addCssClass(element, animation)
+  }
+}
+
+export { animation }
