+++ _jsbuild/home/home-header-animation.js
@@ -1,20 +1,11 @@
 /*  */
-import { scene3D } from './scene-3D'
-import { addCssClass } from '../common/css-class'
+import { animation } from '../common/animation'
 
 const homeHeaderAnimation = () => {
   animation('loader', 'hide')
   animation('center-content', 'show')
   animation('center-content-buttons', 'show')
   animation('down-arrow', 'show')
-  scene3D()
-}
-
-const animation = (elementId, animation) => {
-  const element = document.getElementById(elementId)
-  if (element) {
-    addCssClass(element, animation)
-  }
 }
 
 export { homeHeaderAnimation }
