+++ package.json
@@ -62,7 +62,6 @@
   "dependencies": {
     "bootstrap": "4.3.1",
     "detector-webgl": "2.0.0",
-    "gsap": "2.0.2",
     "intersection-observer": "0.5.1",
     "orbit-controls-es6": "2.0.0",
     "three": "0.100.0",
